+++ app/Http/Controllers/Admin/States/StatesController.php
@@ -6,6 +6,9 @@ use App\Http\Controllers\Controller;
 use App\Models\State\State;
 use Illuminate\Http\Request;
 
+use Illuminate\Database\QueryException;
+
+
 class StatesController extends Controller
 {
     public function __construct(){
@@ -56,7 +59,18 @@ class StatesController extends Controller
 
     public function destroy(State $state)
     {
-        $state->delete();
-        return redirect()->route('admin.states.index')->with('delete', 'El estado se a eliminado correctamente.');
+        // Verifica si el estado es uno de los primeros 6 registros (id del 1 al 6)
+        if ($state->id <= 6) {
+            return redirect()->route('admin.states.index')->with('info', 'Este estado no se puede eliminar ya que es uno de los principales en el sistema');
+        }
+
+        try {
+            $state->delete();
+            return redirect()->route('admin.states.index')->with('delete', 'El estado se ha eliminado correctamente.');
+        } catch (QueryException $e) {
+            if ($e->errorInfo[1] === 1451) {
+                return redirect()->route('admin.states.index')->with('info', 'El estado no se puede eliminar, ya que está relacionado con otros registros.');
+            }
+        }
     }
 }
